+++ src/poly.h
@@ -31,6 +31,6 @@ void poly_set_coefs(Poly *p, const uint32_t *coefs);
 void poly_free(Poly *p);
 
 // n: length of v
-Poly* poly_create_poly_from_coeffs(const uint8_t *v, uint8_t n);
+Poly* poly_create_poly_from_coeffs(const uint16_t *v, uint8_t n);
 
 #endif
